Tested, fixed, and cleaned up.
Added missing IDs to TidyFormInput inputs.

Removed redundant function definitions and consolidated all emphasis and tab management code down to the document mixin level.
